+++ lib/__init__.py
@@ -1,2 +1,2 @@
 # Wazuh Agent Manager Library
-__version__ = "1.6.18"
+__version__ = "1.6.19"